Move tracker to a new file
